Implement reservation

+++ src/main/java/edu/cmu/andrew/workbnb/server/models/Reservation.java
@@ -56,11 +56,11 @@ public class Reservation {
         this.duration = duration;
     }
 
-    public Float getPrice() {
+    public Double getPrice() {
         return price;
     }
 
-    public void setPrice(Float price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
 }
